# swagger.yaml
openapi: "3.0.3"
info:
  title: MySlotMate API
  description: |
    Backend API for **MySlotMate** – a slot-booking platform with wallet-based payments,
    Razorpay payouts, Aadhar (Setu OKYC) verification, AWS S3 file uploads,
    and real-time inbox messaging.

    Every response follows the standard envelope:
    ```json
    { "success": true|false, "data": ..., "message": "...", "error": "..." }
    ```
  version: "2.0.0"
servers:
  - url: http://localhost:{port}
    description: Local development
    variables:
      port:
        default: "8080"
...
paths:
...
  /admin/hosts/applications:
    get:
      tags: [Admin]
      summary: List all pending host applications
      operationId: listPendingApplications
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of pending host applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_HostList"
        "401":
          description: Unauthorized — missing or invalid token
        "403":
          description: Forbidden — caller is not admin
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/hosts/{hostID}/approve:
    post:
      tags: [Admin]
      summary: Approve a host application (status → approved)
      operationId: approveHostApplication
      security:
        - BearerAuth: []
      parameters:
        - name: hostID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Application approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_Host"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Unauthorized — missing or invalid token
        "403":
          description: Forbidden — caller is not admin
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/hosts/{hostID}/reject:
    post:
      tags: [Admin]
      summary: Reject a host application (status → rejected)
      operationId: rejectHostApplication
      security:
        - BearerAuth: []
      parameters:
        - name: hostID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for rejecting the application
                  example: Insufficient experience description
      responses:
        "200":
          description: Application rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_Host"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Unauthorized — missing or invalid token
        "403":
          description: Forbidden — caller is not admin
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Events ──────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: "Firebase ID token passed as `Authorization: Bearer <token>`. Admin endpoints additionally require the token's email to match the `ADMIN_EMAIL` env variable."

  # ── Parameters ──────────────────────────────
...
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ── Schemas ─────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string

    # ── Request Bodies ────────────────────────
...
    Host:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
          nullable: true
        first_name:
          type: string
        last_name:
          type: string
        phn_number:
          type: string
        city:
          type: string
        avatar_url:
          type: string
          nullable: true
        tagline:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        application_status:
          $ref: "#/components/schemas/HostApplicationStatus"
        experience_desc:
          type: string
          nullable: true
        moods:
          type: array
          items:
            type: string
        description:
          type: string
          nullable: true
        preferred_days:
          type: array
          items:
            type: string
        group_size:
          type: integer
          nullable: true
        government_id_url:
          type: string
          nullable: true
        submitted_at:
          type: string
          format: date-time
          nullable: true
        approved_at:
          type: string
          format: date-time
          nullable: true
        rejected_at:
          type: string
          format: date-time
          nullable: true
        is_identity_verified:
          type: boolean
        is_email_verified:
          type: boolean
        is_phone_verified:
          type: boolean
        is_super_host:
          type: boolean
        is_community_champ:
          type: boolean
        expertise_tags:
          type: array
          items:
            type: string
        social_instagram:
          type: string
          nullable: true
        social_linkedin:
          type: string
          nullable: true
        social_website:
          type: string
          nullable: true
        avg_rating:
          type: number
          format: double
          nullable: true
        total_reviews:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    HostApplicationStatus:
      type: string
      enum: [draft, pending, under_review, approved, rejected]
...
    SuccessResponse_Host:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/Host"
    SuccessResponse_HostList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Host"
...